refactor: rename timeout hook as useTimeout
